requirements.txt
Flask==2.3.3
pdfplumber==0.10.3
pandas==2.3.2
openpyxl==3.1.2
xlrd==1.2.0
google-generativeai==0.3.2
nltk==3.8.1
python-dotenv==1.0.0
neo4j==5.25.0